pyproject.toml
[project]
name = "bewise-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "orjson>=3.10.18",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.13",
    "types-pyyaml>=6.0.12.20250516",
]


[tool.ruff]
target-version = "py312"
line-length = 120
lint.select = ["ALL"]
lint.ignore = ["D", "COM812", "TD", "FIX"]

[tool.mypy]
strict = true
